requirements.txt
termcolor
Pillow==6.2.1
PyQt5==5.13.2
PyQt5-sip==12.7.0